django-on-docker/app/requirements.txt
django==3.2.6
Djangorestframework==3.11.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
psycopg2
mysql-connector
pyodbc==4.0.28
requests